RDA: GuardUse is an atom
