+++ angr/analyses/reaching_definitions/atoms.py
@@ -9,6 +9,14 @@ class Atom:
         raise NotImplementedError()
 
 
+class GuardUse(Atom):
+    def __init__(self, target):
+        self.target = target
+
+    def __repr__(self):
+        return '<Guard %#x>' % self.target
+
+
 class Tmp(Atom):
     """
     Represents a variable used by the IR to store intermediate values.
